# backend/internal/api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SwiftInbox External Developer API",
    "version": "1.0.0",
...
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/mailboxes": {
      "post": {
        "tags": ["Mailboxes"],
        "summary": "Create a mailbox",
        "description": "Creates a mailbox for the authenticated developer. The `username` field must be a full mailbox address on the configured mail domain (for example `dev-123@temp.mail.at`).",
        "operationId": "createMailbox",
        "x-usage": { "operation": "mailbox.create", "credits": 1, "category": "mailbox" },
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RequestEmail" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mailbox created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Mailbox" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "get": {
        "tags": ["Mailboxes"],
        "summary": "List mailboxes",
        "description": "Lists all mailboxes owned by the authenticated developer.",
        "operationId": "listMailboxes",
        "x-usage": { "operation": "mailbox.list", "credits": 1, "category": "mailbox" },
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Mailboxes owned by the developer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Mailbox" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/mailboxes/custom": {
      "post": {
        "tags": ["Mailboxes"],
        "summary": "Create a custom mailbox",
        "description": "Creates a mailbox from a username (the local part of the address). The username is sanitized to lowercase alphanumeric characters plus `.`, `-`, and `_` and combined with the configured mail domain. An address that is already in use returns `409 Conflict`.",
        "operationId": "createCustomMailbox",
        "x-usage": { "operation": "mailbox.create", "credits": 1, "category": "mailbox" },
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RequestEmail" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mailbox created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Mailbox" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/mailboxes/{id}": {
      "get": {
        "tags": ["Mailboxes"],
        "summary": "Get a mailbox",
        "description": "Fetches a single mailbox owned by the authenticated developer by its UUID.",
        "operationId": "getMailbox",
        "x-usage": { "operation": "mailbox.list", "credits": 1, "category": "mailbox" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Mailbox UUID.",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "The mailbox.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Mailbox" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "delete": {
        "tags": ["Mailboxes"],
        "summary": "Delete a mailbox by id",
        "description": "Deletes a mailbox owned by the authenticated developer, identified by its UUID. The endpoint also accepts a mailbox address (see `DELETE /api/v1/mailboxes/address/{address}`).",
        "operationId": "deleteMailboxById",
        "x-usage": { "operation": "mailbox.delete", "credits": 1, "category": "mailbox" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Mailbox UUID (or, in practice, a mailbox address).",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Mailbox deleted.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeletedMailbox" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/mailboxes/address/{address}": {
      "delete": {
        "tags": ["Mailboxes"],
        "summary": "Delete a mailbox by address",
        "description": "Deletes a mailbox owned by the authenticated developer, identified by its full address (for example `dev-123@temp.mail.at`).",
        "operationId": "deleteMailboxByAddress",
        "x-usage": { "operation": "mailbox.delete", "credits": 1, "category": "mailbox" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "description": "Full mailbox address.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Mailbox deleted.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeletedMailbox" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/mailboxes/{address}/messages": {
      "get": {
        "tags": ["Messages"],
        "summary": "List messages in a mailbox",
        "description": "Lists the messages in a mailbox reached by address. A developer may read their own mailboxes and any public mailbox; accessing another developer's mailbox returns `403 Forbidden`.",
        "operationId": "listMessages",
        "x-usage": { "operation": "message.list", "credits": 3, "category": "message" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "description": "Mailbox address.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages in the mailbox.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessagesResult" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/mailboxes/{address}/message": {
      "post": {
        "tags": ["Messages"],
        "summary": "List messages in a mailbox (POST)",
        "description": "Alternative POST form of `GET /api/v1/mailboxes/{address}/messages`. Takes no request body.",
        "operationId": "listMessagesPost",
        "x-usage": { "operation": "message.list", "credits": 3, "category": "message" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "description": "Mailbox address.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages in the mailbox.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessagesResult" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/messages/{id}": {
      "get": {
        "tags": ["Messages"],
        "summary": "Get a message",
        "description": "Fetches a single message body by UUID, including parsed text, HTML, and attachment metadata.",
        "operationId": "getMessage",
        "x-usage": { "operation": "message.get", "credits": 3, "category": "message" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Message UUID.",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "The message.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageDetail" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/message/{id}": {
      "post": {
        "tags": ["Messages"],
        "summary": "Get a message (POST)",
        "description": "Alternative POST form of `GET /api/v1/messages/{id}`. Takes no request body.",
        "operationId": "getMessagePost",
        "x-usage": { "operation": "message.get", "credits": 3, "category": "message" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Message UUID.",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "The message.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageDetail" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/messages/{id}/attachment/{index}": {
      "get": {
        "tags": ["Attachments"],
        "summary": "Download an attachment",
        "description": "Downloads a single attachment of a message by zero-based index. The response is the raw attachment bytes; its `Content-Type` is the attachment's content type (or `application/octet-stream` when unknown) and `Content-Disposition` is set to `inline` or `attachment`.",
        "operationId": "getAttachment",
        "x-usage": { "operation": "attachment.get", "credits": 2, "category": "message" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Message UUID.",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "index",
            "in": "path",
            "required": true,
            "description": "Zero-based attachment index.",
            "schema": { "type": "integer", "minimum": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "The raw attachment bytes.",
            "content": {
              "application/octet-stream": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/webhooks": {
      "get": {
        "tags": ["Webhooks"],
        "summary": "List webhooks",
        "description": "Lists all webhooks owned by the authenticated developer.",
        "operationId": "listWebhooks",
        "x-usage": { "operation": "webhook.get", "credits": 1, "category": "api" },
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "The developer's webhooks.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Webhook" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "post": {
        "tags": ["Webhooks"],
        "summary": "Create a webhook",
        "description": "Creates a webhook and links it to one or more owned mailboxes. The returned webhook includes the one-time `secret` used to verify signatures in the `X-Webhook-Signature` header.",
        "operationId": "createWebhook",
        "x-usage": { "operation": "webhook.create", "credits": 5, "category": "api" },
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WebhookCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Webhook" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/webhooks/create": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Create a webhook (POST /create)",
        "description": "Alias of `POST /api/v1/webhooks` that accepts the same request body and returns the same response.",
        "operationId": "createWebhookAlias",
        "x-usage": { "operation": "webhook.create", "credits": 5, "category": "api" },
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WebhookCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Webhook" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/webhooks/dead-letters": {
      "get": {
        "tags": ["Webhooks"],
        "summary": "List dead letters",
        "description": "Lists undelivered webhook deliveries (dead letters) that exhausted retries for the authenticated developer.",
        "operationId": "listWebhookDeadLetters",
        "x-usage": { "operation": "webhook.dead_letters", "credits": 1, "category": "api" },
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "The developer's dead letters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/DeadLetter" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/webhooks/dead-letters/{id}/seen": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Mark a dead letter as seen",
        "description": "Marks a dead letter as seen so it no longer counts as unread.",
        "operationId": "markWebhookDeadLetterSeen",
        "x-usage": { "operation": "webhook.dead_letter_seen", "credits": 1, "category": "api" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Dead letter UUID.",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Dead letter marked as seen.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeadLetterSeen" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
    "/api/v1/webhooks/{id}/mailboxes": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Link a mailbox to a webhook",
        "description": "Links a mailbox owned by the developer to a webhook so events are delivered for it. Linking a mailbox that is already linked returns `409 Conflict`.",
        "operationId": "linkWebhookMailbox",
        "x-usage": { "operation": "webhook.create", "credits": 5, "category": "api" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Webhook UUID.",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WebhookMailboxRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mailbox linked to the webhook.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WebhookMailboxResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "dev_<64 hex chars>",
        "description": "API key sent as a Bearer token in the `Authorization` header (the middleware also accepts the legacy `Authentication` header):\n\n```\nAuthorization: Bearer dev_<64 hex characters>\n```\n\nCreate or manage keys in the developer dashboard (**Api keys** page)."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request is missing or malformed (invalid path value, invalid JSON body, wrong domain, invalid expiry, etc.).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
...
      "RequestEmail": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "For `POST /api/v1/mailboxes` this must be a full mailbox address on the configured mail domain (for example `dev-123@temp.mail.at`). For `POST /api/v1/mailboxes/custom` this is the local part (for example `dev-123`), which is sanitized and combined with the mail domain."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Optional expiry timestamp. Must be in the future; omitted to use the default mailbox lifetime."
          }
        },
        "required": ["username"]
      },
      "Mailbox": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Mailbox id. Present on get/list responses; not returned when a mailbox is created."
          },
          "address": { "type": "string", "description": "Full mailbox address." },
          "createdAt": { "type": "string", "format": "date-time" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "createdBy": {
            "type": ["string", "null"],
            "format": "uuid",
            "description": "Developer id that created the mailbox, or null for public mailboxes."
          }
        },
        "required": ["address", "createdAt", "expiresAt", "createdBy"]
      },
...
    }
  }
}